math_index_add() now returns flushed payload size.
writer tested for indexing 200 * 7 expression.

+++ math-index-v3/math-index.h
@@ -57,7 +57,7 @@ math_index_lookup(math_index_t, const char *);
 
 void math_index_print(math_index_t);
 
-int math_index_add(math_index_t, doc_id_t, exp_id_t, struct subpaths);
+size_t math_index_add(math_index_t, doc_id_t, exp_id_t, struct subpaths);
 
 /* utilities */
 int    mk_path_str(struct subpath*, int, char*);
